+++ Buzz/package.json
@@ -10,10 +10,13 @@
     "test": "jest"
   },
   "dependencies": {
+    "@react-navigation/native": "^6.1.6",
+    "@react-navigation/native-stack": "^6.9.12",
     "react": "18.2.0",
     "react-native": "0.71.7",
     "react-native-paper": "^5.7.2",
-    "react-native-safe-area-context": "^4.5.2",
+    "react-native-safe-area-context": "^4.5.3",
+    "react-native-screens": "^3.20.0",
     "react-native-sqlite-storage": "^6.0.1",
     "react-native-vector-icons": "^9.2.0"
   },
